==> 4_recommender_model.ipynb <==
### Pre-installs

# %pip install pandas
# %pip install numpy
# %pip install matplotlib
# %pip install seaborn
# %pip install sklearn
# %pip install -U scikit-learn
# %pip install spotipy

import random
import json
import pandas as pd
import numpy as np
import seaborn as sns
import matplotlib.pyplot as plt
import sklearn
from sklearn.preprocessing import MinMaxScaler
from sklearn.metrics.pairwise import cosine_similarity
import ast
import spotipy
from spotipy.oauth2 import SpotifyClientCredentials
from spotipy.oauth2 import SpotifyOAuth

### Inputs

with open("data/variables.txt", "r") as file:
    data = json.load(file)

input_sentence = data['input_sentence']
input_playlist = data['input_playlist']
recc_num = data['recc_num']
client_id = data['client_id']
client_secret = data['client_secret']
sp_dc = data['sp_dc']
sp_key = data['sp_key']

### CSV & Spotify Information

# Importing the datasets
dataset = pd.read_csv("data/tracks_with_genres_lang_emotion.csv", index_col=0)
df = pd.read_csv("data/tracks.csv")

# Creating the dataframe
df_generated = pd.DataFrame(dataset)
df_generated['genres'] = df_generated['genres'].apply(ast.literal_eval)

df_generated.iloc[:,3:19].head()

# Spotify API Authentication Information
client_credentials_manager = SpotifyClientCredentials(client_id=client_id, client_secret=client_secret)
sp = spotipy.Spotify(auth_manager=SpotifyOAuth(client_id=client_id,
                                               client_secret=client_secret,
                                               redirect_uri="http://localhost/",
                                               scope="playlist-modify-private",
                                               show_dialog=True,
                                               cache_path="token.txt"
                                               ))

# Read the file as an array
emotions_prob = np.loadtxt("data/emotions_prob.txt")
emotions_prob = emotions_prob

# Print the array
# print(emotions_prob)

with open("data/nlp_genre.txt", "r") as file:
    nlp_genre_data = file.read()
# nlp_genre_data

with open("data/nlp_language.txt", "r") as file:
    nlp_language_data = file.read()
# nlp_language_data


### Processing the Input Song

Determine the y_kmeans cluster that the input wants.

highest_index = np.argmax(emotions_prob)
# highest_index

df_generated.groupby('y_kmeans')[['danceability', 'energy', 'key', 'loudness', 'mode',
       'speechiness', 'acousticness', 'instrumentalness', 'liveness',
       'valence', 'tempo', 'time_signature']].mean()

Cluster 0 is Happy Songs <br>
Cluster 1 is Sad Songs <br>
Cluster 2 is Angry/Tense Songs

df_generated.groupby('y_kmeans')[['danceability', 'energy', 'key', 'loudness', 'mode',
       'speechiness', 'acousticness', 'instrumentalness', 'liveness',
       'valence', 'tempo', 'time_signature']].min()

df_generated.groupby('y_kmeans')[['danceability', 'energy', 'key', 'loudness', 'mode',
       'speechiness', 'acousticness', 'instrumentalness', 'liveness',
       'valence', 'tempo', 'time_signature']].max()

proportion_increment = (emotions_prob[highest_index] - 0.33) / (1 - 0.33)
# print(proportion_increment)

Happy:

Positive correlation with:<br>
Valence (high valence indicates positivity)<br>
Energy (high energy indicates upbeat and lively music)<br>
Danceability (high danceability indicates music suitable for dancing)<br><br>
Negative correlation with:<br>
None (for the purpose of this categorization, as features associated with happiness generally have positive correlations)

Tense/Angry:

Positive correlation with:<br>
Energy (high energy indicates intensity and aggression)<br>
Loudness (high loudness indicates intensity)<br>
Acousticness (high acousticness for rawness and intensity)<br>
Instrumentalness (high instrumentalness for intensity and aggression)<br>
Liveness (high liveness for raw energy)<br>
Speechiness (high speechiness for aggression or urgency)<br><br>
Negative correlation with:<br>
Valence (low valence indicates negativity)<br>
Tempo (potentially, as slower tempos might indicate brooding or aggression)

Sad:

Positive correlation with:<br>
Acousticness (high acousticness for emotional depth and vulnerability)<br>
Instrumentalness (high instrumentalness for emotional expressiveness)<br>
Liveness (high liveness for authenticity and vulnerability)<br>
Speechiness (high speechiness for introspection or storytelling)<br><br>
Negative correlation with:<br>
Valence (low valence indicates sadness)<br>
Energy (low energy indicates calmness and introspection)<br>
Loudness (potentially, as softer dynamics may convey melancholy)

# Create a blank dataframe with columns
df_blank = pd.DataFrame({'danceability': [0], 'energy': [0], 'key': [0], 'loudness': [0], 'mode': [0], 'speechiness': [0], 
                         'acousticness': [0], 'instrumentalness': [0], 'liveness': [0], 'valence': [0], 'tempo': [0]})
# df_blank

# Declare the relationship between the features and the emotion clusters.
selected_cluster = df_generated.groupby('y_kmeans').get_group(highest_index)

feature_tweak_dict = {'danceability': ['+', '=', '='],
                      'energy': ['+', '-', '+'], 
                      'loudness': ['-', '+', '-'], 
                      'speechiness': ['+', '+', '+'], 
                      'acousticness': ['+', '+', '-'],
                      'instrumentalness': ['+', '+', '+'], 
                      'liveness': ['+', '+', '+'], 
                      'valence': ['+', '-', '-'], 
                      'tempo': ['+', '=', '-']
                    }

# Function to compute the value of the input based on the relationship to the features.
def compute_value(col, proportion_increment, i, highest_index):
  if (feature_tweak_dict[i][highest_index] == '+'):
    return col.min() + ((col.max() - col.min()) * proportion_increment)
  elif (feature_tweak_dict[i][highest_index] == '-'):
    return col.min() + ((col.max() - col.min()) * (1 - proportion_increment))
  else:
    return col.mean()

feature_tweak_dict['danceability'][highest_index]

for i in feature_tweak_dict.keys():
  df_blank[i] = compute_value(selected_cluster[i], proportion_increment, i, highest_index)
# df_blank

Key determination:

Happy:

Major keys are often associated with happier or uplifting emotions.
Keys such as C major, D major, E major, and G major are commonly perceived as happy or joyful.
In our scaled representation, these keys would correspond to the numbers 0, 2, 4, and 7.

Tense/Angry:

Minor keys or keys with dissonant intervals may evoke feelings of tension or aggression.
Keys such as D minor, E minor, G minor, and A minor are often associated with darker or more intense emotions.
Additionally, certain keys with tritones or augmented intervals, such as F# major/Gb major or C# minor/Db minor, can also convey tension.
In our scaled representation, these keys would correspond to the numbers 2, 4, 7, 9, and potentially others.

Sad:

Minor keys are commonly associated with sadness or melancholy.
Keys such as A minor, C minor, D minor, and F minor are often perceived as somber or melancholic.
Additionally, certain keys with flat keys such as Eb major/D# major or Ab major/G# major may also convey sadness.
In our scaled representation, these keys would correspond to the numbers 9, 0, 2, 5, and potentially others.
It's important to note that these associations are generalizations and may not hold true for all musical contexts or individual interpretations. Additionally, cultural and personal experiences can heavily influence how different keys are perceived emotionally.

In summary:

Happy:

C Major / A Minor: 0<br>
D Major / B Minor: 2<br>
E Major / C# Minor: 4<br>
G Major / E Minor: 7

Tense/Angry:

D# Major / C# Minor: 1<br>
F Major / D Minor: 5<br>
G# Major / F Minor: 8<br>
A# Major / G Minor: 10

Sad:

C# Major / A# Minor: 1<br>
E♭ Major / C Minor: 3<br>
F# Major / D# Minor: 6<br>
A♭ Major / F Minor: 8

For our recommender, we will match the cluster to the range of keys, and randomly select one for the input instance.

emotion_key_dict = {0: ['0', '2', '4', '7'], 1: ['1', '3', '6', '8'], 2: ['1', '5', '8', '10']}
chosen_keys = emotion_key_dict[highest_index]

random_index = random.randint(0, len(chosen_keys)-1)
random_key = int(chosen_keys[random_index])
df_blank['key'] = random_key
# random_key

# Mode
random_mode = random.randint(0, 1)
df_blank['mode'] = random_mode
# random_mode

df_blank['supersetted_genres'] = nlp_genre_data
df_blank['supersetted_genres'][0] = ast.literal_eval(df_blank['supersetted_genres'][0])
df_blank['language'] = nlp_language_data
df_blank['y_kmeans'] = highest_index
# df_blank

df_blank['language'] = df_blank['language'].apply(lambda row: 'zh' if (row == 'z1') or (row == 'zh-cn') or (row == 'zh-tw') else row)

### Dataset processing

# Counting the number of genres in for each song.
def count_items_in_list(lst):
    return len(lst)

# Apply the function to the DataFrame column
df_generated['genres_count'] = df_generated['genres'].apply(count_items_in_list)
df_generated.head()

# print(df_generated['id'].count())
df_generated = df_generated[~df_generated['name'].str.contains('remaster', case=False)]
# print(df_generated['id'].count())

# print(df_generated.shape)
new_df = df_generated[(df_generated['genres_count'] != 0)]
# print(new_df.shape)

new_df['language'] = new_df['language'].apply(lambda row: 'zh' if (row == 'z1') or (row == 'zh-cn') or (row == 'zh-tw') else row)

new_df = new_df.drop(new_df[(new_df['language'] == 'm7') | (new_df['language'] == 'iw')].index)

### Supersetting genres

# Determining the frequency of the genre. Giving the input song's genre priority to go first.
genre_extract = new_df['genres'].reset_index()
genre_wo_index = genre_extract['genres']
tally = genre_wo_index.count()
genre_dict = {}

for i in range (tally):
    input = genre_wo_index.loc[i]
    if len(input) > 0:
        for n in input:
            if n in genre_dict.keys():
                genre_dict[n] += 1
            else:
                genre_dict[n] = 1

# print(genre_dict)

genres_df = pd.Series(genre_dict)
# genres_df.head()

# Creating the genre supersets, with priority to the input song's genre.
# Since the input song is the first one to be run, its genre will be able to start the classification of other sub-genres under it.

# This can be further improved. Allowing for later genres to take over earlier genres as the superset if it more general.
# E.g.: Detroit Hip Hop is found before general hip hop. Currently they will be seperated into 2 seperate genres because the first word 'detroit' is
# not 'hip hop' and is thus overlooked.

def generate_superset_mapping(genre_dict):
    superset_mapping = {}

    for genre, frequency in genre_dict.items():
        subset_added = False
        for superset in list(superset_mapping.keys()):
            if genre in superset:
                superset_mapping[genre] = [genre]
                superset_mapping[genre].extend(superset_mapping[superset])
                del superset_mapping[superset]
                subset_added = True
                break
            elif superset in genre:
                if superset in superset_mapping:
                    superset_mapping[superset].append(genre)
                    subset_added = True
                    break
        if not subset_added:
            superset_mapping[genre] = [genre]

    return superset_mapping

# Generate superset mapping
superset_mapping = generate_superset_mapping(genre_dict)

# # Output the superset mapping
# for superset, subsets in superset_mapping.items():
#     print(f"{superset}: {subsets}")


# Now with the superset mapped out, apply it onto the all the songs to determine their supersetted genre(s).
# This can be improved. Instead of only returning the first superset genre, it should return multiple in an array.
new_df['supersetted_genres'] = new_df['genres'].apply(lambda genres: [m for n in genres for m in superset_mapping if n in m])

### Combining the dataset

merged_df = new_df[['danceability', 'energy', 'key', 'loudness', 'mode',
       'speechiness', 'acousticness', 'instrumentalness', 'liveness',
       'valence', 'tempo', 'time_signature', 'genres', 'language', 'supersetted_genres', 'y_kmeans']]
merged_df = pd.concat([df_blank, merged_df])

### Encoding

# Determining how many keys there are.
merged_df.groupby('key').count()

# One-hot encoding the keys.
key_encoded = pd.get_dummies(merged_df['key'], columns=[i for i in range(12)], prefix='key').astype(int)

key_encoded.head()

# Adding keys encoding into the dataset.
df_2 = pd.concat([merged_df, key_encoded], axis=1)

df_2.reset_index(drop=True, inplace=True)

# One-hot Encoding of all the superset genres of each song in the dataframe.

genres_dummies = pd.get_dummies(merged_df['supersetted_genres'].apply(pd.Series).stack(), prefix="genre").groupby(level=0).sum()

# Concatenate the binary columns with the original DataFrame
df_2 = pd.concat([df_2, genres_dummies], axis=1)

# Filling in any null values. (In case)
df_2 = df_2.fillna(0)
df_2.head()

### Weighting

numericals = {'danceability': 0, 'energy': 0.15, 'loudness': 0.1, 'mode': 0.10, 'speechiness': 0.05, 'acousticness': 0.20, 
    'instrumentalness': 0.25, 'liveness': 0.05, 'valence': 0.15, 'tempo': 0.1}
categoricals = {'key_': 0.05, 'genre_': 0.15}
numericals_key = numericals.keys()
categoricals_key = categoricals.keys()

from sklearn.preprocessing import MinMaxScaler
# Normalize numerical features
numerical_features = ['loudness', 'tempo']
scaler = MinMaxScaler()
df_2[numerical_features] = scaler.fit_transform(df_2[numerical_features])

# # Standardize Year
# df_2['standardized_year'] = scaler.fit_transform(df_2[['year']])

for i in numericals_key:
    df_2[i] = df_2[i] * numericals[i]

for n in categoricals_key:
    for z in df_2.columns:
        if n in z:
            df_2[z] = df_2[z] * float(categoricals[n])

### Cosine Similarity & Output Extraction

# Feature selection, removing non-useful rows.
def processing(df_2):
    # Create a feature matrix
    feature_matrix = df_2.drop(['supersetted_genres', 'language', 'y_kmeans', 'time_signature', 'genres'], axis=1)

    # Apply normalization to the entire feature matrix
    feature_matrix = scaler.fit_transform(feature_matrix)

    # This is the input song. We are isolating it from the dataframe first.
    input_song = feature_matrix[0]
    
    arr = []

    # Returns comparison value of first song in dataset to all others in an array.
    # DataFrame can be customized before Feature Selection to change which song the subject of comparison should be.
    for i in range(1, len(feature_matrix)):
        # Take the i-th song from the feature matrix
        compare_song = feature_matrix[i]
        # Perform cosine similarity.
        cosine_sim = cosine_similarity([input_song], [compare_song])
        cs_value = cosine_sim[0][0]  # Take the single similarity value between the input song and the i-th song.
        arr.append(cs_value)

    return arr

arr = processing(df_2)
# arr

def compile_suggestions(arr):
    suggestions = []
    
    # Compiling all song ids that are above the pre-determined threshold for 'like the first song'.
    for i in range(len(arr)):
        if arr[i] > 0.6:
            suggestions.append(i)
    
    # print(suggestions)
    return suggestions
suggestions = compile_suggestions(arr)

# Extracting out the end result information.
df_main = new_df[new_df.index.isin(suggestions)]

df_main = df_main[['id', 'name', 'release_date', 'language', 'y_kmeans']]
df_main.head()

id_artist_mapping = df.set_index('id')['artists'].to_dict()

# Add 'artist' column to df1 using the mapping
df_main['artists'] = df_main['id'].map(id_artist_mapping)
df_main.head()

### Hard Category Filtering

# Keeping only the songs that match the input song's language.
df_filter = pd.DataFrame()

# Filter rows where the genre is present in the 'supersetted_genres' column
language_filter = df_main[df_main['language'].apply(lambda x: nlp_language_data == x)]
# Concatenate filtered DataFrame with df_filter
df_filter = pd.concat([df_filter, language_filter])
df_filter.shape

df_filter = df_filter[df_filter['y_kmeans'] == highest_index]

# print(df_filter.shape)
df_filter.drop_duplicates(subset=["id"], keep='first', inplace=True)
# print(df_filter.shape)

df_filter = df_filter[['id', 'name', 'artists', 'release_date']]
df_filter.head()

### Recommendations Output

# Generating recc_num of the matched songs at random.
def recommend(df_filter, recc_num):
    recs = []
    recommended_artists = set()  # Track recommended artists
    if len(df_filter) == 0:
        print("Sorry! There are no songs similar enough to " + input_sentence + "!")
    elif len(df_filter) < recc_num:
        recc_num = len(df_filter)
        print("Uh oh! We only found " + str(recc_num) + " recommendations!")
    print('Here are your recommendations!')
    while len(recs) < recc_num:
        recc = df_filter.sample().to_dict()
        artist = list(recc['artists'].values())[0]
        recommended_artists.add(artist)
        name = list(recc['name'].values())[0]
        artists = artist[1:len(artist)-1]  # Extract artist name from brackets
        release_date = str(list(recc['release_date'].values())[0])
        id = str(list(recc['id'].values())[0])
        print(str(len(recs) + 1) + '. ' + name + ' by ' + artists + ' published on ' + release_date)
        df_filter = df_filter.drop(list(recc['id'])[0])
        recs.append((id, name, release_date))
    return recs

# Call the function with your DataFrame and the desired number of recommendations
recs_for_user = recommend(df_filter, recc_num)

user_id = sp.current_user()["id"]
song_uris = []
for song in recs_for_user:
    result = sp.track(song[0])
    print(result)
    try:
        uri = result["uri"]
        song_uris.append(uri)
    except IndexError:
        print(f"{song[0]} doesn't exist in Spotify. Skipped.")

playlist = sp.user_playlist_create(user=user_id, name=input_playlist, public=False)
sp.playlist_add_items(playlist_id=playlist["id"], items=song_uris)
